pyproject.toml
[tool.poetry]
name = "mlnext"
version = "1.0.0"
description = "Machine learning utilities for Tensorflow/Keras."
authors = ["Gorden Platz <gplatz@phoenixcontact.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy="^1.19.5"
pydantic="^1.8.1"
pandas= "^1.2.3"
scipy= "^1.6.1"
PyYAML= "^5.4.1"
scikit_learn= "^0.24.1"
matplotlib="^3.3.4"
pydot="^1.4.2"
tensorflow="^2.4.1"

[tool.poetry.dev-dependencies]
testfixtures= "^6.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
